Move test_object_detector_export_coreml_with_nms to its own file with the appropriate protobuf workarounds
