+++ test/unity/toolkits/CMakeLists.txt
@@ -8,6 +8,7 @@ add_subdirectory(sparse_similarity)
 add_subdirectory(pattern_mining)
 add_subdirectory(neural_net)
 add_subdirectory(object_detection)
+add_subdirectory(coreml_export)
 
 make_boost_test(kmeans_test.cxx
   REQUIRES unity_clustering util numerics
